requirements.txt
fastapi==0.103.2
uvicorn[standard]==0.23.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
pydantic==2.4.2
pydantic-settings==2.0.3
requests==2.31.0
python-dotenv==1.0.0
bcrypt==4.0.1
passlib==1.7.4
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
httpx==0.25.0
flake8==6.1.0
black==23.10.1
isort==5.12.0
